refactor: PlanService 패키지 경로 변경 및 트랜잭션 적용
1. PlanService 클래스의 패키지를 application에서 application.service로 이동한다.
2. findPlanByType 메서드에 @Transactional(readOnly = true)를 추가하여 읽기 전용 트랜잭션을 적용한다.
3. 관련된 import 경로를 수정하여 변경 사항을 반영한다.

Resolves #22
